refactor: fix build errors
