+++ packages/console/src/components/AuditLogTable/index.tsx
@@ -1,5 +1,5 @@
 import { LogResult } from '@logto/schemas';
-import type { LogDto } from '@logto/schemas/lib/types/log-legacy.js';
+import type { LogDto } from '@logto/schemas/lib/types/log-legacy';
 import { conditional, conditionalString } from '@silverhand/essentials';
 import classNames from 'classnames';
 import { useTranslation } from 'react-i18next';
