+++ packages/console/src/pages/AuditLogDetails/index.tsx
@@ -1,5 +1,5 @@
 import type { User } from '@logto/schemas';
-import type { LogDto } from '@logto/schemas/lib/types/log-legacy.js';
+import type { LogDto } from '@logto/schemas/lib/types/log-legacy';
 import classNames from 'classnames';
 import { useTranslation } from 'react-i18next';
 import { useLocation, useParams } from 'react-router-dom';
