+++ packages/core/src/include.d/import-meta.d.ts
@@ -1,0 +1,10 @@
+interface ImportMeta {
+  jest: typeof jest & {
+    // Almost same as `jest.mock()`, but factory is required
+    unstable_mockModule: <T = unknown>(
+      moduleName: string,
+      factory: () => T,
+      options?: jest.MockOptions
+    ) => typeof jest;
+  };
+}
