+++ packages/core/src/middleware/koa-auth.ts
@@ -1,7 +1,6 @@
 import type { IncomingHttpHeaders } from 'http';
 
-import { UserRole } from '@logto/schemas';
-import { managementResource } from '@logto/schemas';
+import { UserRole, managementResource } from '@logto/schemas';
 import type { Optional } from '@silverhand/essentials';
 import { conditional } from '@silverhand/essentials';
 import { jwtVerify } from 'jose';
