+++ packages/core/src/routes/phrase.test.ts
@@ -3,10 +3,9 @@ import type { SignInExperience } from '@logto/schemas';
 import { adminConsoleApplicationId, adminConsoleSignInExperience } from '@logto/schemas';
 import { mockEsm, mockEsmWithActual, pickDefault } from '@logto/shared/esm';
 
-import { createMockProvider } from '#src/test-utils/oidc-provider.js';
-
 import { zhCnTag } from '#src/__mocks__/custom-phrase.js';
 import { mockSignInExperience } from '#src/__mocks__/index.js';
+import { createMockProvider } from '#src/test-utils/oidc-provider.js';
 
 const { jest } = import.meta;
