+++ packages/core/src/routes/session/middleware/koa-guard-session-action.ts
@@ -1,5 +1,4 @@
-import { SignInMode } from '@logto/schemas';
-import { adminConsoleApplicationId } from '@logto/schemas';
+import { SignInMode, adminConsoleApplicationId } from '@logto/schemas';
 import type { MiddlewareType } from 'koa';
 import type { Provider } from 'oidc-provider';
 import { errors } from 'oidc-provider';
