+++ packages/core/src/routes/session/password.ts
@@ -1,6 +1,5 @@
 import { passwordRegEx, usernameRegEx } from '@logto/core-kit';
-import { SignInIdentifier, UserRole } from '@logto/schemas';
-import { adminConsoleApplicationId } from '@logto/schemas';
+import { SignInIdentifier, UserRole, adminConsoleApplicationId } from '@logto/schemas';
 import type { Provider } from 'oidc-provider';
 import { object, string } from 'zod';
