+++ packages/integration-tests/src/tests/api/session.test.ts
@@ -1,5 +1,4 @@
-import { SignInIdentifier } from '@logto/schemas';
-import { adminConsoleApplicationId } from '@logto/schemas';
+import { SignInIdentifier, adminConsoleApplicationId } from '@logto/schemas';
 import { assert } from '@silverhand/essentials';
 
 import {
