+++ packages/schemas/alterations/1.0.0_beta.10-1664265197-custom-phrases.ts
@@ -1,6 +1,6 @@
 import { sql } from 'slonik';
 
-import type { AlterationScript } from '../lib/src/types/alteration.js';
+import type { AlterationScript } from '../src/types/alteration.js';
 
 const alteration: AlterationScript = {
   up: async (pool) => {
