+++ packages/schemas/alterations/1.0.0_beta.14-1667900481-add-passcode-type-continue.ts
@@ -1,6 +1,6 @@
 import { sql } from 'slonik';
 
-import type { AlterationScript } from '../lib/src/types/alteration.js';
+import type { AlterationScript } from '../src/types/alteration.js';
 
 const alteration: AlterationScript = {
   up: async (pool) => {
