+++ packages/schemas/alterations/next-1671039448-add-user-name-index.ts
@@ -1,6 +1,6 @@
 import { sql } from 'slonik';
 
-import type { AlterationScript } from '../lib/src/types/alteration.js';
+import type { AlterationScript } from '../src/types/alteration.js';
 
 const alteration: AlterationScript = {
   up: async (pool) => {
