+++ packages/schemas/alterations/next-1671080370-terms-of-use.ts
@@ -1,7 +1,7 @@
 import type { DatabaseTransactionConnection } from 'slonik';
 import { sql } from 'slonik';
 
-import type { AlterationScript } from '../lib/src/types/alteration.js';
+import type { AlterationScript } from '../src/types/alteration.js';
 
 type DeprecatedTermsOfUse = {
   enabled: boolean;
