+++ packages/schemas/tsconfig.json
@@ -8,7 +8,6 @@
   },
   "include": [
     "src",
-    "src-models",
     "alterations"
   ]
 }
